functions/image-processor/requirements.txt
pillow==9.5.0
numpy==1.24.3
rembg==2.0.50
opencv-python==4.7.0.72
opencv-contrib-python==4.7.0.72
